Inbetween sync with real networks

+++ PyAlgorithms/src/results.py
@@ -1,17 +1,28 @@
 import os
 
-results_path = os.path.join("..", "results.dat")
+
+def collect_synthetic_results():
+    collect_results("synthetic")
 
 
-def collect_results():
-    data_path = os.path.join("..", "results.dat")
+def collect_real_results():
+    collect_results("real")
+
+
+def collect_results(dir_name):
     missing_names = []
+    data_path = os.path.join("..", dir_name + "_results.dat")
+
+    # First run there will be no results to gather
+    if not os.path.exists(data_path):
+        print("There are no results yet")
+        return
 
     # Open file where the results will be written into
     with open(data_path, "w") as dat_file:
         dat_file.write("config_name, demon_nmi, demon_omega, demon_nf1, oslom_nmi, oslom_omega, oslom_nf1, wnw_nmi, wnw_omega, wnw_nf1\n")
 
-        network_dir = os.path.join("..", "networks")
+        network_dir = os.path.join("..", dir_name + "_networks")
         network_names = [name for name in os.listdir(network_dir) if os.path.isdir(os.path.join(network_dir, name))]
 
         # Each network name
